Attempt to move logic of endlessScroll
