+++ src/articles/OwnArticles.js
@@ -3,73 +3,39 @@ import useShadowRef from "../hooks/useShadowRef";
 import LoadingAnimation from "../components/LoadingAnimation";
 import { setTitle } from "../assorted/setTitle";
 import strings from "../i18n/definitions";
-import { getPixelsFromScrollBarToEnd } from "../utils/misc/getScrollLocation";
 
 import ArticlePreview from "./ArticlePreview";
 
 import SortingButtons from "./SortingButtons";
 
 import * as s from "../components/TopMessage.sc";
+import useEndlessScrolling from "../hooks/useEndlessScrolling";
+import { ADD_ACTIONS } from "../utils/endlessScrolling/add_actions";
 
 export default function OwnArticles({ api }) {
   const [articleList, setArticleList] = useState(null);
   const [originalList, setOriginalList] = useState(null);
-  const [currentPage, setCurrentPage] = useState(0);
-  const [isWaitingForNewArticles, setIsWaitingForNewArticles] = useState(false);
-  const [noMoreArticlesToShow, setNoMoreArticlesToShow] = useState(false);
 
-  const articleListRef = useShadowRef(articleList);
-  const currentPageRef = useShadowRef(currentPage);
-  const noMoreArticlesToShowRef = useShadowRef(noMoreArticlesToShow);
-  const isWaitingForNewArticlesRef = useShadowRef(isWaitingForNewArticles);
-
-  function insertNewArticlesIntoArticleList(
-    fetchedArticles,
-    newCurrentPage,
-    newArticles,
-  ) {
-    if (fetchedArticles.length === 0) {
-      setNoMoreArticlesToShow(true);
-    }
-    newArticles = newArticles.concat(fetchedArticles);
-    setArticleList(newArticles);
-    setCurrentPage(newCurrentPage);
-    setIsWaitingForNewArticles(false);
-  }
-
-  function handleScroll() {
-    let scrollBarPixelDistToPageEnd = getPixelsFromScrollBarToEnd();
-    let articlesHaveBeenFetched =
-      currentPageRef.current !== undefined &&
-      articleListRef.current !== undefined;
-
-    if (
-      scrollBarPixelDistToPageEnd <= 50 &&
-      !isWaitingForNewArticlesRef.current &&
-      !noMoreArticlesToShowRef.current &&
-      articlesHaveBeenFetched
-    ) {
-      setIsWaitingForNewArticles(true);
-      setTitle("Getting more articles...");
-
-      let newCurrentPage = currentPageRef.current + 1;
-      let newArticles = [...articleListRef.current];
-      api.getSavedUserArticles(newCurrentPage, (articles) => {
-        insertNewArticlesIntoArticleList(articles, newCurrentPage, newArticles);
-        setTitle("Saved Articles");
-      });
-    }
-  }
+  const [handleScroll, isWaitingForNewArticles, noMoreArticlesToShow] =
+    useEndlessScrolling(api, articleList, setArticleList, "Saved Articles");
 
   useEffect(() => {
     setTitle("Saved Articles");
-    api.getSavedUserArticles(currentPage, (articles) => {
+    api.getSavedUserArticles(0, (articles) => {
       setArticleList(articles);
       setOriginalList(articles);
     });
-    window.addEventListener("scroll", handleScroll, true);
+    window.addEventListener(
+      "scroll",
+      handleScroll(ADD_ACTIONS.ADD_SAVED_ARTICLES),
+      true,
+    );
     return () => {
-      window.removeEventListener("scroll", handleScroll, true);
+      window.removeEventListener(
+        "scroll",
+        handleScroll(ADD_ACTIONS.ADD_SAVED_ARTICLES),
+        true,
+      );
     };
   }, []);
