+++ src/utils/endlessScrolling/add_actions.js
@@ -1,0 +1,7 @@
+const ADD_ACTIONS = Object.freeze({
+  ADD_SAVED_ARTICLES: 1,
+  ADD_RECOMMENDED_ARTICLES: 2,
+  ADD_SEARCH_ARTICLES: 3,
+});
+
+export { ADD_ACTIONS };
